requirements.txt
streamlit
pandas
yfinance
plotly.express
